+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Re-added `values`  in `Intent` and `Intents`.
 * Added support for inline replies and `MessageBehavior#reply` to quickly create a reply to a message. #110
 
+## Changes
+
+* Gateway reconnect retries reset on handshake. #68
+
 # 0.7.0-M1
 
 ## Changes